fixed mistakes

+++ e04/src/main/java/myapp/StudentRecord.java
@@ -1,6 +1,6 @@
 package myapp;
 
 
-public record StudentRecord(int id, String name, int grade){
+public record StudentRecord(int id, String name, double grade){
     
 }
